requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
antlr4-python3-runtime==4.8
asgiref==3.4.1; python_version >= '3.6'
attrs==21.2.0
click==8.0.3
django==3.2.9
hydra-core==1.1.1
omegaconf==2.1.1
pytz==2021.3
pyyaml==6.0; python_version >= '3.6'
sqlparse==0.4.2; python_version >= '3.5'
